Fixed previously diabled tests
